Restaurantes/requirements.txt
Django==1.9.4
django-mongodb-engine==0.6.0
django-registration-redux==1.4
djangorestframework==3.3.3
djangotoolbox==1.8.0
mongoengine==0.10.6
pymongo==3.1.1
